+++ pgmpy/parameter/adapter/SkproAdapter.py
